move to NodeFinder namespace
